\name{SpatialCondition-package}
\alias{SpatialCondition-package}
\alias{SpatialCondition}
\docType{package}
\title{
Package for estimating spatial and spatiotemporal variation in individual condition using Gaussian random fields and optionally with covariates
}
\description{
This package is intended estimating spatial and spatiotemporal variation in individual condition using Gaussian random fields and optionally with covariate.
}
\details{
\tabular{ll}{
Package: \tab SpatialCondition\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-30\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
